+++ lib/src/flexible_bottom_sheet_route.dart
@@ -23,8 +23,13 @@ const Duration _bottomSheetDuration = Duration(milliseconds: 500);
 /// [minHeight] - min height in fractional value for bottom sheet. e.g. 0.1.
 /// [initHeight] - init height in fractional value for bottom sheet. e.g. 0.5.
 /// [maxHeight] - init height in fractional value for bottom sheet. e.g. 0.5.
+/// [isCollapsible] - will the bottom sheet collapse.
+/// [isDismissible] - the bottom sheet will be dismissed when user taps on the scrim.
+/// [isExpand] - whether the widget should expand to fill the available space in its parent or not.
 /// [isModal] - if true, overlay background with dark color.
 /// [anchors] - list of sizes in fractional value that the bottom sheet can accept.
+/// [keyboardBarrierColor] - keyboard color.
+/// [duration] - animation speed when opening bottom sheet.
 Future<T?> showFlexibleBottomSheet<T>({
   required BuildContext context,
   required FlexibleDraggableScrollableWidgetBuilder builder,
@@ -38,6 +43,7 @@ Future<T?> showFlexibleBottomSheet<T>({
   bool isModal = true,
   List<double>? anchors,
   Color? keyboardBarrierColor,
+  Duration? duration,
 }) {
   assert(debugCheckHasMediaQuery(context));
   assert(debugCheckHasMaterialLocalizations(context));
@@ -56,6 +62,7 @@ Future<T?> showFlexibleBottomSheet<T>({
       isModal: isModal,
       anchors: anchors,
       keyboardBarrierColor: keyboardBarrierColor,
+      duration: duration,
     ),
   );
 }
@@ -77,6 +84,8 @@ Future<T?> showFlexibleBottomSheet<T>({
 /// [headerHeight] - head size.
 /// Set both [minHeaderHeight] and [maxHeaderHeight].
 /// Set one ([maxHeaderHeight] or [headerHeight]).
+/// [keyboardBarrierColor] - keyboard color.
+/// [duration] - animation speed when opening bottom sheet.
 Future<T?> showStickyFlexibleBottomSheet<T>({
   required BuildContext context,
   required FlexibleDraggableScrollableHeaderWidgetBuilder headerBuilder,
@@ -95,6 +104,7 @@ Future<T?> showStickyFlexibleBottomSheet<T>({
   double? headerHeight,
   Decoration? decoration,
   Color? keyboardBarrierColor,
+  Duration? duration,
 }) {
   assert(maxHeaderHeight != null || headerHeight != null);
   assert(debugCheckHasMediaQuery(context));
@@ -118,6 +128,7 @@ Future<T?> showStickyFlexibleBottomSheet<T>({
       maxHeaderHeight: maxHeaderHeight ?? headerHeight!,
       decoration: decoration,
       keyboardBarrierColor: keyboardBarrierColor,
+      duration: duration,
     ),
   );
 }
@@ -140,12 +151,13 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
   final Decoration? decoration;
   final ThemeData? theme;
   final Color? keyboardBarrierColor;
+  final Duration? duration;
 
   @override
   final String? barrierLabel;
 
   @override
-  Duration get transitionDuration => _bottomSheetDuration;
+  Duration get transitionDuration => duration ?? _bottomSheetDuration;
 
   @override
   bool get barrierDismissible => isDismissible;
@@ -173,6 +185,7 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
     this.maxHeaderHeight,
     this.decoration,
     this.keyboardBarrierColor,
+    this.duration,
     RouteSettings? settings,
   }) : super(settings: settings);